provider/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
drf-spectacular[sidecar]==0.26.5
drf_social_oauth2==2.1.3
gunicorn==20.1.0
djoser==2.1.0